Adequa nomenclaturas ao padrão do react e reorganiza arquivos divididos por views
View para desktop e mobile
