--- src/containers/app_bar/app_bar.js
@@ -1,47 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import Typography from 'material-ui/Typography';
-import Button from 'material-ui/Button';
-import IconButton from 'material-ui/IconButton';
-import MiniDrawer from '../../containers/menu_lateral/menu_lateral.js';
-
-// é para ser um container
-
-const styles = {
-  root: {
-    flexGrow: 1,
-  },
-  flex: {
-    //flex: 1
-  },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20,
-  },
-};
-
-function ButtonAppBar(props) {
-  const { classes } = props;
-  return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton className={classes.menuButton} color="inherit" aria-label="Menu">
-          </IconButton>
-          <Typography variant="title" color="inherit" className={classes.flex}>
-            Quem me Representa?
-          </Typography>
-        </Toolbar>
-      </AppBar>
-    </div>
-  );
-}
-
-ButtonAppBar.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(ButtonAppBar);
