+++ src/desktop-view/components/votacao/ExpansionPanel.js
