+++ src/desktop-view/containers/menu_lateral/MenuSections.js
