+++ src/desktop-view/containers/qmr_na_midia/QMRNaMidiaContainer.js
