+++ src/mobile-view/components/votacao/VotacaoMobile.js
