+++ src/mobile-view/containers/main/MainMobileContainer.js
