+++ src/mobile-view/containers/votacoes/VotacoesMobileContainer.js
